refactor(vector_float.(cpp|hpp)): migrate to compat/db_vector
