+++ src/object/object_primitive.c
@@ -58,7 +58,7 @@
 #endif /* !defined (SERVER_MODE) */
 
 #include "dbtype.h"
-#include "vector_float.hpp"
+#include "db_vector.hpp"
 #include "memory_private_allocator.hpp"
 // XXX: SHOULD BE THE LAST INCLUDE HEADER
 #include "memory_wrapper.hpp"
