--- src/object/vector_float.hpp
@@ -1,31 +1,0 @@
-/*
- * Copyright 2008 Search Solution Corporation
- * Copyright 2016 CUBRID Corporation
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-/*
- * vector_float.hpp - Utility structures and functions used for vector float
- */
-
-#ifndef _VECTOR_FLOAT_HPP_
-#define _VECTOR_FLOAT_HPP_
-
-#include "dbtype_def.h"
-#include <string>
-
-std::string db_vector_float_to_string (const DB_VECTOR_FLOAT &vf);
-
-#endif // _VECTOR_FLOAT_HPP_
